+++ docs/sync.js
@@ -36,7 +36,7 @@
 
 /**
  * This describes the different options used to create a {@link Realm} instance with Realm Cloud synchronization.
- * @typedef {Object} Realm.Sync~SyncConfiguration
+ * @typedef {Object} Realm.App.Sync~SyncConfiguration
  * @property {Realm.User} user - A {@link Realm.User} object obtained by calling `Realm.App.logIn`.
  * @property {string|number|BSON.ObjectId|null} partitionValue - The value of the partition key.
  * @property {function} [error] - A callback function which is called in error situations.
@@ -44,21 +44,21 @@
  *    `category`, and `code`.
  *
  * @property {Object} [customHttpHeaders] - A map (string, string) of custom HTTP headers.
- * @property {Realm.Sync~OpenRealmBehaviorConfiguration} [newRealmFileBehavior] - Whether to create a new file and sync in background or wait for the file to be synced.
+ * @property {Realm.App.Sync~OpenRealmBehaviorConfiguration} [newRealmFileBehavior] - Whether to create a new file and sync in background or wait for the file to be synced.
        If not set, the Realm will be downloaded before opened.
- * @property {Realm.Sync~OpenRealmBehaviorConfiguration} [existingRealmFileBehavior] - Whether to open existing file and sync in background or wait for the sync of the
+ * @property {Realm.App.Sync~OpenRealmBehaviorConfiguration} [existingRealmFileBehavior] - Whether to open existing file and sync in background or wait for the sync of the
  *    file to complete and then open. If not set, the Realm will be downloaded before opened.
  */
 
 /**
  * Specify how to open a synced Realm.
  *
- * @typedef {Object} Realm.Sync~OpenRealmBehaviorConfiguration
+ * @typedef {Object} Realm.App.Sync~OpenRealmBehaviorConfiguration
  * @property {string} type - how to open a Realm - 'downloadBeforeOpen' to wait for download to complete or 'openImmediately' to open the local Realm
  * @property {number} [timeOut] - how long to wait for a download (in ms). Default: infinity
  * @property {string} [timeOutBehavior] - what to do when download times out - 'openLocalRealm' to open the local Realm or 'throwException' to throw an exception.
- * @see {@link Realm.Sync~openLocalRealmBehavior}
- * @see {@link Realm.Sync~downloadBeforeOpenBehavior}
+ * @see {@link Realm.App.Sync~openLocalRealmBehavior}
+ * @see {@link Realm.App.Sync~downloadBeforeOpenBehavior}
  */
 
 /**
@@ -66,13 +66,13 @@
  * If this is the first time you open the Realm, it will be empty while the server data is being downloaded
  * in the background.
  *
- * @typedef {Realm.Sync~OpenRealmBehaviorConfiguration} Realm.Sync~openLocalRealmBehavior
+ * @typedef {Realm.App.Sync~OpenRealmBehaviorConfiguration} Realm.App.Sync~openLocalRealmBehavior
  */
 
 /**
  * The default behavior settings if you want to wait for downloading a synchronized Realm to complete before opening it.
  *
- * @typedef {Realm.Sync~OpenRealmBehaviorConfiguration} Realm.Sync~downloadBeforeOpenBehavior
+ * @typedef {Realm.App.Sync~OpenRealmBehaviorConfiguration} Realm.App.Sync~downloadBeforeOpenBehavior
  */
 
  /**
@@ -145,6 +145,16 @@
      * @type {Realm.Auth.EmailPasswordAuth}
      */
     get emailPasswordAuth() { }
+
+    /**
+     * Returns an instance of an app from the cache. If an app with the specified id
+     * hasn't been created, null is returned.
+     *
+     * @param {string} appId
+     * @returns {Realm.App|null}
+     * @since v10.0.0
+     */
+    getCachedApp(appId) { }
  }
 
 
@@ -158,42 +168,48 @@
 class Sync {
 
     /**
-     * Calling this method will force Realm to attempt to reconnect to the server immediately.
+     * Calling this method will force Realm to attempt to reconnect the Realm App to the server immediately.
      *
      * Realm will reconnect automatically, but by using exponential backoff. This means that if the device is offline for
      * a long time, restoring the connection after it comes back online can take longer than expected. In situations
      * where it is possible to detect the network condition (e.g. Airplane mode). Manually calling this method can
      * provide a smoother user experience.
+     *
+     * @param {Realm.App} app - The Realm app.
      */
-    static reconnect() { }
+    static reconnect(app) { }
 
     /**
      * Set the sync log level. You can only set the log level once, and you must do it after creating an App instance
      * but before opening any Realms.
+     *
+     * @param {Realm.App} app - The Realm app.
+     * @param {Realm.Sync~LogLevel} level - The new log level
      * @example
      * {
      * const app = new Realm.App(getAppConfig());
-     * Realm.Sync.setLogLevel("all");
+     * Realm.App.Sync.setLogLevel("all");
      * const user = await app.logIn(credentials);
      * const realm = await Realm.open(getRealmConfig(user));
      * }
-     * @param {Realm.Sync~LogLevel} level - The log level.
+     * @param {Realm.App.Sync~LogLevel} level - The log level.
      */
-    static setLogLevel(level) { }
+    static setLogLevel(app, level) { }
 
     /**
-     * Enable multiplexing multiple sync sessions over a single connection.
+     * Enable multiplexing multiple sync sessions over a single connection for a Realm app.
      * When having a lot of synchronized realms open the system might run out of file
      * descriptors because of all the open sockets to the server. Session multiplexing
      * is designed to alleviate that, but it might not work with a server configured with
      * fail-over. Only use if you're seeing errors about reaching the file descriptor limit
      * and you know you are using many sync sessions.
+     * @param {Realm.App} app - The Realm app.
      */
-    static enableSessionMultiplexing() { }
+    static enableSessionMultiplexing(app) { }
 
     /**
-     * A callback passed to `Realm.Sync.setLogger` when instrumenting the Realm Sync client with a custom logger.
-     * @callback Realm.Sync~logCallback
+     * A callback passed to `Realm.App.Sync.setLogger` when instrumenting the Realm Sync client with a custom logger.
+     * @callback Realm.App.Sync~logCallback
      * @param {number} level The level of the log entry between 0 and 8 inclusively.
      * Use this as an index into `['all', 'trace', 'debug', 'detail', 'info', 'warn', 'error', 'fatal', 'off']` to get the name of the level.
      * @param {string} message The message of the log entry.
@@ -202,17 +218,19 @@ class Sync {
     /**
      * Capture the sync client's log. You can only set the log level once, and you must do it after creating an App instance
      * but before opening any Realms.
+     *
+     * @param {Realm.App} app - the Realm app.
+     * @param {Realm.Sync~logCallback} logger - The log callback.
      * @example
      * {
      * const app = new Realm.App(getAppConfig());
-     * Realm.Sync.setLogger((level, message) => console.log(`[${level}] ${message}`);
+     * Realm.App.Sync.setLogger((level, message) => console.log(`[${level}] ${message}`);
      * const user = await app.logIn(credentials);
      * const realm = await Realm.open(getRealmConfig(user));
      * }
-     * @param {Realm.Sync~logCallback} logger - The log callback.
-     * @see {Realm.Sync~setLogLevel}
+     * @see {Realm.App.Sync~setLogLevel}
      */
-    static setLogger(logger) { }
+    static setLogger(app, logger) { }
 
     /**
      * Set the application part of the User-Agent string that will be sent to the Realm Object Server when a session
@@ -220,13 +238,15 @@ class Sync {
      *
      * This method can only be called up to the point where the first Realm is opened. After that, the User-Agent
      * can no longer be changed.
+     * @param {Realm.App} the Realm app
      * @param {string} the user agent description
      */
-    static setUserAgent(userAgent) { }
+    static setUserAgent(app, userAgent) { }
 
     /**
      * Initiate a client reset. The Realm must be closed prior to the reset.
      *
+     * @param {Realm.App} [app] - The app where the Realm was opened.
      * @param {string} [path] - The path to the Realm to reset.
      * Throws error if reset is not possible.
      * @example
@@ -234,29 +254,49 @@ class Sync {
      *   const config = { sync: { user, partitionValue } };
      *   config.sync.error = (sender, error) => {
      *     if (error.name === 'ClientReset') {
-     *       Realm.Sync.initiateClientReset(original_path);
+     *       Realm.Sync.initiateClientReset(app, original_path);
      *       // copy required objects from Realm at error.config.path
      *     }
      *   }
      * }
      */
-    static initiateClientReset(path) { }
+    static initiateClientReset(app, path) { }
 
     /**
      * Returns `true` if Realm still has a reference to any sync sessions regardless of their state.
      * If `false` is returned it means that no sessions currently exist.
+     * @param {Realm.App} [app] - The app where the Realm was opened.
      */
-    static _hasExistingSessions() { }
+    static _hasExistingSessions(app) { }
+
+    /**
+     * Returns all sync sessions for a user.
+     *
+     * @param {Realm.User} user  - the user.
+     * @returns {Array<Realm.App.Sync.Session>} an array of sessions
+     * @since 10.0.0
+     */
+    static getAllSyncSessions(user) { }
+
+    /**
+     * Returns the session associated with a user and partition value.
+     *
+     * @param {Realm.User} user
+     * @param {string|number|ObjectId|null} partitionValue
+     * @returns {Realm.App.Sync.Session} the session
+     * @since 10.0.0
+     */
+    static getSyncSession(user, partitionValue) { }
 }
 
 /**
- * @typedef Realm.Sync~LogLevel
+ * @typedef Realm.App.Sync~LogLevel
  * @type {("all"|"trace"|"debug"|"detail"|"info"|"warn"|"error"|"fatal"|"off")}
  */
 
 /**
  * Class that describes authentication errors in the Realm Object Server
- * @memberof Realm.Sync
+ * @memberof Realm.App.Sync
  */
 class AuthError extends Error {
     /**
@@ -274,7 +314,7 @@ class AuthError extends Error {
 
 /**
  * Describes an error when an incompatible synced Realm is opened. The old version of the Realm can be accessed in readonly mode using the configuration() member
- * @memberof Realm.Sync
+ * @memberof Realm.App.Sync
  */
 class IncompatibleSyncedRealmError {
     /**
@@ -325,11 +365,11 @@ class Credentials {
     static anonymous() { }
 
     /**
-     * Creates credentials with a custom provider and user identifier.
+     * Creates credentials with a JSON Web Token (JWT) provider and user identifier.
      * @param {string} token A string identifying the user. Usually an identity token or a username.
      * @return {Credentials} An instance of `Credentials` that can be used in {@linkcode Realm.App.logIn}.
      */
-    static custom(token) { }
+    static jwt(token) { }
 
     /**
      * Creates credentials with a MongoDB Realm function and user identifier.
@@ -508,6 +548,25 @@ class User {
     get id() { }
 
     /**
+     * Gets an array of identities for this user on MongoDB Realm Cloud.
+     * Each element in the array is an object with properties userId and providerType.
+     * @type {Array<Object>}
+     */
+    get identities() { }
+
+    /**
+     * Gets the provider type for the identity.
+     * @type {string}
+     */
+    get providerType() { }
+
+    /**
+     * Gets the device id. `null` if no device id.
+     * @type {string}
+     */
+    get deviceId() { }
+
+    /**
      * Gets this user's access token. This is the user's credential for accessing the MongoDB
      * Realm Cloud and should be treated as sensitive data.
      * @type {string}
@@ -604,15 +663,15 @@ class User {
      * Returns a connection to the MongoDB service.
      *
      * @example
-     * let blueWidgets = user.remoteMongoClient('myClusterName')
+     * let blueWidgets = user.mongoClient('myClusterName')
      *                       .db('myDb')
      *                       .collection('widgets')
      *                       .find({color: 'blue'});
      *
      * @param {string} serviceName
-     * @returns {Realm~RemoteMongoDB}
+     * @returns {Realm~MongoDB}
      */
-    remoteMongoClient(serviceName) { }
+    mongoClient(serviceName) { }
 
     /**
      * @class Realm.User~Push Access to the operations of the push service.
@@ -647,7 +706,7 @@ class User {
  * client (and a local Realm file on disk), and the server (and a remote Realm at a given URL stored on a Realm Object Server).
  * Sessions are always created by the SDK and vended out through various APIs. The lifespans of sessions
  * associated with Realms are managed automatically.
- * @memberof Realm.Sync
+ * @memberof Realm.App.Sync
  */
 class Session {
     /**
@@ -724,9 +783,9 @@ class Session {
      * connection. In that case, any connection change is sent to all sessions.
      *
      * Can be either:
-     *  - Realm.Sync.ConnectionState.Disconnected: No connection to the server is available.
-     *  - Realm.Sync.ConnectionState.Connecting: An attempt to connect to the server is in progress.
-     *  - Realm.Sync.ConnectionState.Connected: The connection to the server is active and data can be synchronized.
+     *  - Realm.App.Sync.ConnectionState.Disconnected: No connection to the server is available.
+     *  - Realm.App.Sync.ConnectionState.Connecting: An attempt to connect to the server is in progress.
+     *  - Realm.App.Sync.ConnectionState.Connected: The connection to the server is active and data can be synchronized.
      *
      * Data will only be synchronized with the Realm ObjectServer if this method returns `Connected` and `state()`
      * returns `Active` or `Dying`.
@@ -800,7 +859,7 @@ class Session {
  *    or the listener being added. The virtual path (i.e. the portion of the
  *    URL after the protocol and hostname) is passed as an argument.
  *  * `'change'`: Emitted whenever the data within a Realm matching the filter
- *    regex has changed. A [ChangeEvent]{@link Realm.Sync.ChangeEvent} argument
+ *    regex has changed. A [ChangeEvent]{@link Realm.App.Sync.ChangeEvent} argument
  *    is passed containing information about which Realm changed and what
  *    objects within the Realm changed.
  *  * `'delete'`: Emitted whenever a Realm matching the filter regex has been
@@ -829,7 +888,7 @@ class Session {
  * module.exports = {onchange, oncavailable, ondelete};
  *
  * // server script
- * Realm.Sync.addListener(realmServerURL, adminUser, '.*', new Realm.Worker('my-worker'));
+ * Realm.App.Sync.addListener(realmServerURL, adminUser, '.*', new Realm.Worker('my-worker'));
  *
  * @memberof Realm
  */
@@ -851,29 +910,41 @@ class Worker {
 
 
 /**
- * The RemoteMongoDB service can be used to get database and collection objects for interacting with MongoDB data.
- * @alias Realm~RemoteMongoDB
+ * The MongoDB service can be used to get database and collection objects for interacting with MongoDB data.
+ * @alias Realm~MongoDB
  */
-class RemoteMongoDB {
+class MongoDB {
+    /**
+     * Get the service name.
+     * @return {string} The service name.
+     */
+    get serviceName() { }
+
     /**
      * Get the interface to a remote MongoDB database.
      *
      * @param {string} databaseName The name of the database.
-     * @returns {Realm~RemoteMongoDBDatabase} The remote MongoDB database.
+     * @returns {Realm~MongoDBDatabase} The remote MongoDB database.
      */
     db(databaseName) { }
 }
 
 /**
- * The RemoteMongoDB service can be used to get database and collection objects for interacting with MongoDB data.
- * @alias Realm~RemoteMongoDBDatabase
+ * The MongoDB service can be used to get database and collection objects for interacting with MongoDB data.
+ * @alias Realm~MongoDBDatabase
  */
-class RemoteMongoDBDatabase {
+class MongoDBDatabase {
+    /**
+     * Get the database name.
+     * @return {string} The database name.
+     */
+    get name() { }
+
     /**
      * Get the interface to a remote MongoDB collection.
      *
      * @param {string} name The name of the collection.
-     * @returns {Realm.RemoteMongoDBCollection} The remote MongoDB collection.
+     * @returns {Realm.MongoDBCollection} The remote MongoDB collection.
      */
     collection(name) { }
 }
@@ -882,7 +953,13 @@ class RemoteMongoDBDatabase {
  * A remote collection of documents in a MongoDB database.
  * @memberof Realm
  */
-class RemoteMongoDBCollection {
+class MongoDBCollection {
+    /**
+     * Gets the name of the collection.
+     * @return {string} The name.
+     */
+    get name() { }
+
     /**
      * Finds the documents which match the provided query.
      *
@@ -973,7 +1050,7 @@ class RemoteMongoDBCollection {
     async count(filter, options) { }
 
     /**
-     * @typedef Realm.RemoteMongoDBCollection~InsertOneResult Result of inserting a document
+     * @typedef Realm.MongoDBCollection~InsertOneResult Result of inserting a document
      * @property insertedId The id of the inserted document
      */
 
@@ -982,12 +1059,12 @@ class RemoteMongoDBCollection {
      * Note: If the document is missing an _id, one will be generated for it by the server.
      *
      * @param {object} document The document.
-     * @returns {Promise<Realm.RemoteMongoDBCollection~InsertOneResult>} The _id of the inserted document.
+     * @returns {Promise<Realm.MongoDBCollection~InsertOneResult>} The _id of the inserted document.
      */
     async insertOne(document) { }
 
     /**
-     * @typedef Realm.RemoteMongoDBCollection~InsertManyResult Result of inserting many documents
+     * @typedef Realm.MongoDBCollection~InsertManyResult Result of inserting many documents
      * @property {Array} insertedIds The ids of the inserted documents
      */
 
@@ -996,12 +1073,12 @@ class RemoteMongoDBCollection {
      * If any values are missing identifiers, they will be generated by the server.
      *
      * @param {object[]} documents The array of documents.
-     * @returns {Promise<Realm.RemoteMongoDBCollection~InsertManyResult>} The _ids of the inserted documents.
+     * @returns {Promise<Realm.MongoDBCollection~InsertManyResult>} The _ids of the inserted documents.
      */
     async insertMany(documents) { }
 
     /**
-     * @typedef {object} Realm.RemoteMongoDBCollection~DeleteResult Result of deleting documents
+     * @typedef {object} Realm.MongoDBCollection~DeleteResult Result of deleting documents
      * @property {number} deletedCount The number of documents that were deleted.
      */
 
@@ -1009,7 +1086,7 @@ class RemoteMongoDBCollection {
      * Deletes a single matching document from the collection.
      *
      * @param {object} filter A filter applied to narrow down the result.
-     * @returns {Promise<Realm.RemoteMongoDBCollection~DeleteResult>}
+     * @returns {Promise<Realm.MongoDBCollection~DeleteResult>}
      */
     async deleteOne(filter) { }
 
@@ -1017,12 +1094,12 @@ class RemoteMongoDBCollection {
      * Deletes multiple documents.
      *
      * @param {object} filter A filter applied to narrow down the result.
-     * @returns {Promise<Realm.RemoteMongoDBCollection~DeleteResult>}
+     * @returns {Promise<Realm.MongoDBCollection~DeleteResult>}
      */
     async deleteMany(filter) { }
 
     /**
-     * @typedef {object} Realm.RemoteMongoDBCollection~UpdateResult Result of updating documents
+     * @typedef {object} Realm.MongoDBCollection~UpdateResult Result of updating documents
      * @property {number} matchedCount The number of documents that matched the filter.
      * @property {number} modifedCount The number of documents matched by the query.
      * @property [upsertedId] The identifier of the inserted document if an upsert took place.
@@ -1036,7 +1113,7 @@ class RemoteMongoDBCollection {
      * @param {object} [options] Additional options to apply.
      * @param {boolean} [options.upsert=false] if true, indicates that MongoDB should insert a new document that matches the
      * query filter when the query does not match any existing documents in the collection.
-     * @returns {Promise<Realm.RemoteMongoDBCollection~UpdateResult>}
+     * @returns {Promise<Realm.MongoDBCollection~UpdateResult>}
      */
     async updateOne(filter, update, options) { }
 
@@ -1048,12 +1125,12 @@ class RemoteMongoDBCollection {
      * @param {object} [options] Additional options to apply.
      * @param {boolean} [options.upsert=false] if true, indicates that MongoDB should insert a new document that matches the
      * query filter when the query does not match any existing documents in the collection.
-     * @returns {Promise<Realm.RemoteMongoDBCollection~UpdateResult>}
+     * @returns {Promise<Realm.MongoDBCollection~UpdateResult>}
      */
     async updateMany(filter, update, options) { }
 
     /**
-     * @typedef {object} Realm.RemoteMongoDBCollection~ChangeEvent An event in a change stream.
+     * @typedef {object} Realm.MongoDBCollection~ChangeEvent An event in a change stream.
      *
      * Note that which properties are present will depend on both the
      * `operationType` field, which is itself always present, and the MongoDB
@@ -1091,7 +1168,7 @@ class RemoteMongoDBCollection {
      * @param {object} [options.filter] A filter for which change events you are interested in.
      * @param {any[]} [options.ids] A list of ids that you are interested in watching
      *
-     * @yields {Realm.RemoteMongoDBCollection~ChangeEvent} a change event
+     * @yields {Realm.MongoDBCollection~ChangeEvent} a change event
      */
     async* watch(options) {}
 }