+++ packages/realm-web/src/tests/OAuth2Helper.test.ts
@@ -26,38 +26,37 @@ describe("OAuth2Helper", () => {
     it("can initiate a flow", async () => {
         const windowsOpened: URL[] = [];
         const storage = new MemoryStorage();
-        const helper = new OAuth2Helper(
-            storage,
-            async () => "http://localhost:1337",
-            urlString => {
-                const url = new URL(urlString);
-                windowsOpened.push(url);
-                // Simulating another tab updating the storage
-                setTimeout(() => {
-                    const state = url.searchParams.get("state");
-                    storage.set(
-                        `oauth2:state(${state}):result`,
-                        JSON.stringify({
-                            appId: "default-app-id",
-                            userAuth: "our-little-secret",
-                        }),
-                    );
-                }, 0);
-                return null;
-            },
-        );
+        const helper = new OAuth2Helper(storage, urlString => {
+            const url = new URL(urlString);
+            windowsOpened.push(url);
+            // Simulating another tab updating the storage
+            setTimeout(() => {
+                const state = url.searchParams.get("state");
+                storage.set(
+                    `oauth2:state(${state}):result`,
+                    JSON.stringify({
+                        appId: "default-app-id",
+                        userAuth: "our-little-secret",
+                    }),
+                );
+            }, 0);
+            return null;
+        });
 
         const credentials = Credentials.google<
             Realm.Credentials.OAuth2RedirectPayload
         >("http://localhost:1337/callback");
         expect(typeof credentials.payload.redirectUrl).equals("string");
 
-        const result = await helper.initiate(credentials);
+        const state = helper.generateState();
+        const redirectUrl = `https://some-external-service.com?state=${state}`;
+        const result = await helper.openWindowAndWaitForRedirect(
+            redirectUrl,
+            state,
+        );
         expect(windowsOpened.length).equals(1);
         const [url] = windowsOpened;
-        expect(url.hostname).equals("localhost");
-        expect(url.port).equals("1337");
-        expect(url.pathname).equals("/auth/providers/oauth2-google/login");
+        expect(url.hostname).equals("some-external-service.com");
         expect(result.userAuth).equals("our-little-secret");
     });
 });